fix(conductor): prevent TTS goroutine leak on StreamLLM failure, warningMinutes returns int
Replace ttsLoop for-range with select to respect context cancellation,
preventing goroutine leak when StreamLLM fails before sentCh is closed.
Add fanOut.Close() on StreamLLM error path so TTSAccumulator goroutine
is always cleaned up. Change warningMinutes to return int, removing
redundant casts at call sites. Fix test numbering gap in session_ws_test.go.

+++ internal/interview/conductor_unit_test.go
@@ -11,7 +11,7 @@ func TestWarningMinutes(t *testing.T) {
 	tests := []struct {
 		name     string
 		duration time.Duration
-		want     float64
+		want     int
 	}{
 		{"45 min session", 45 * time.Minute, 5},  // round(45/9)=5
 		{"30 min session", 30 * time.Minute, 3},  // round(30/9)=3
